Say what a sample weight means instead of reading it off the family (#362)
`sample_weight` used to mean whatever the family said it meant — EDM prior
weights for Tweedie, replication counts for everything else, with no way to say
otherwise. It is now a declared parameter, and the default for every family is
the prior reading.

    SuperGLM(..., weight_semantics="prior")      # new default, every family
    SuperGLM(..., weight_semantics="frequency")  # the previous non-Tweedie behaviour

The prior reading is R's and glum's single weight argument, statsmodels'
`var_weights` and Stata's `aweight`; the frequency reading is statsmodels'
`freq_weights` and Stata's `fweight`. They coincide only at `w == 1` — not at
integer `w` — and at fractional `w` only the prior reading is a likelihood at
all, because a row cannot be replicated 0.4 times. Exposure is continuous, so
the previous default was wrong for this library's primary use case, silently,
in every weighted non-Tweedie fit.

The contract reaches the dispersion likelihood size and residual d.f.; the
Gaussian, Gamma and Tweedie REML scale profilers; learned knot and bin
geometry; the reported log-likelihood's normalizer and hence AIC and BIC; the
randomized quantile residuals; the negative-binomial theta profile; the Q-Q
envelope's simulation; Vuong; released-fit reporting; the sklearn wrappers; and
feature-subset clones. All of them now read one value resolved once, and the
helpers that decide it lost their family argument so no site can silently keep
the old rule.

**The `"frequency"` arm reproduces the previous release bit for bit** — phi,
deviance, effective_df, log-likelihood, AIC, BIC, beta, predictions, lambda and
the resolved knot vector, on both `fit` and `fit_reml`, verified against a
detached `origin/master` worktree.

Measured behaviour change on the new default (Gamma, 1500 rows, sum(w)=1850.7):
residual d.f. 1835.9 -> 1485.2, phi x1.2358, every Wald standard error x1.1117,
REML lambda x1.3345. AIC and BIC are not comparable across contracts — their
~736-point fall is entirely the likelihood's (y, w)-only normalizer.

Counting families have a support: `w Y ~ Poisson(w mu)` and `NB2(w mu, w theta)`
live on the non-negative integers, and the binomial prior form is normalised
only at `w == 1`. Off-lattice fits emit `PriorWeightLatticeWarning` rather than
being refused, because coefficients, fitted means and deviance are unaffected
(the contracts share a score equation) except under `theta="auto"`, where the
profiled theta enters the variance — which the warning says.

`estimate_p` refuses `weight_semantics="frequency"` with non-unit weights rather
than profiling the Tweedie power under the wrong likelihood; the combination was
unreachable before, so nothing regressed.

Performance: the prior Gamma REML arm is 1.28x the frequency arm at 50k rows
with continuous weights, down from 1.67x after profiling found duplicate profile
calls, a discarded digamma pass inside `polygamma`, and array bookkeeping.

Reviewed across four rounds; 29 findings, all fixed.
`docs/development/migrations/weight-semantics-prior.md` carries the full table.

Closes #349.

+++ src/superglm/reml/gradient.py
@@ -106,7 +106,13 @@ def reml_direct_hessian(
     penalty_ranks: dict[str, float] | None = None,
     penalty_caches: dict | None = None,
     pirls_result: PIRLSResult | None = None,
-    n_obs: int = 0,
+    # The declared weight contract's likelihood size, not the physical row
+    # count -- `sum(w)` under `"frequency"`, the positive-row count under
+    # `"prior"`. It is the `n` in this Hessian's `0.5 * (n - M_p) * log(D)`
+    # scale term, and it must match the `n` the OBJECTIVE uses for that same
+    # term or the Newton step is inconsistent with the surface it steps on.
+    # Float, because `sum(w)` is not an integer.
+    n_obs: float = 0.0,
     phi_hat: float = 1.0,
     penalty_nullity: float | None = None,
     dH_extra: Mapping[int, NDArray | CompactSymmetricOperator] | None = None,
